Plassert søkefunksjon i egen komponent og plassert søke funksjonalitet i searchFunction
